+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.  
 The format is based on [Keep a Changelog](https://keepachangelog.com/),  
 and this project adheres to [Semantic Versioning](https://semver.org/).
+
+## [0.11.0]- 2025-01-08
+### Added
+- Imperial or Metric management is determined by SimBrief or the app itself, with SimBrief taking priority.
+
+### Bugfix
+- Cargo time not correct
+- Issues with Secret and SimBriefID
+
 ## [0.10.3.1] - 2025-01-07
 ### Bugfix
 - simbriefID null value